Begin conversion to "Old" incident related classes
