--- Assets/Resources/IncidentData/Old Age Death.json
@@ -1,82 +1,0 @@
-{
-  "$type": "System.Collections.Generic.List`1[[Game.Incidents.CoreIncident, Assembly-CSharp]], mscorlib",
-  "$values": [
-    {
-      "$id": "1",
-      "$type": "Game.Incidents.CoreIncident, Assembly-CSharp",
-      "incidentName": "Old Age Death",
-      "tags": {
-        "$type": "System.Collections.Generic.List`1[[Game.Incidents.IIncidentTag, Assembly-CSharp]], mscorlib",
-        "$values": [
-          {
-            "$id": "2",
-            "$type": "Game.Incidents.WorldTag, Assembly-CSharp",
-            "tags": {
-              "$type": "System.Collections.Generic.List`1[[Game.Enums.WorldTagType, Assembly-CSharp]], mscorlib",
-              "$values": [
-                0
-              ]
-            }
-          },
-          {
-            "$id": "3",
-            "$type": "Game.Incidents.SpecialCaseTag, Assembly-CSharp",
-            "tagType": 0
-          },
-          {
-            "$id": "4",
-            "$type": "Game.Incidents.InstigatorTag, Assembly-CSharp",
-            "type": "Game.WorldGeneration.World, Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null"
-          }
-        ]
-      },
-      "weight": 100,
-      "required": {
-        "$type": "System.Collections.Generic.List`1[[Game.Incidents.IncidentModifier, Assembly-CSharp]], mscorlib",
-        "$values": [
-          {
-            "$id": "5",
-            "$type": "Game.Incidents.DeathModifier, Assembly-CSharp",
-            "parent": {
-              "$ref": "1"
-            },
-            "replaceID": 0,
-            "Creatures": null,
-            "Incidents": {
-              "$type": "System.Collections.Generic.List`1[[Game.Incidents.IncidentModifier, Assembly-CSharp]], mscorlib",
-              "$values": []
-            }
-          },
-          {
-            "$id": "6",
-            "$type": "Game.Incidents.OldAgeModifier, Assembly-CSharp",
-            "parent": {
-              "$ref": "1"
-            },
-            "replaceID": 0,
-            "Incidents": {
-              "$type": "System.Collections.Generic.List`1[[Game.Incidents.IncidentModifier, Assembly-CSharp]], mscorlib",
-              "$values": []
-            }
-          }
-        ]
-      },
-      "optional": {
-        "$type": "System.Collections.Generic.List`1[[Game.Incidents.IncidentModifier, Assembly-CSharp]], mscorlib",
-        "$values": []
-      },
-      "context": null,
-      "Incidents": {
-        "$type": "System.Collections.Generic.List`1[[Game.Incidents.IncidentModifier, Assembly-CSharp]], mscorlib",
-        "$values": [
-          {
-            "$ref": "5"
-          },
-          {
-            "$ref": "6"
-          }
-        ]
-      }
-    }
-  ]
-}
