--- Assets/Resources/IncidentData/Old Age Death.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a13b961202165b44eb4bd689fc35b4e5
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
