--- Assets/Resources/IncidentData/Test Death.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c764d07d2f822e047845f0993359b342
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
