+++ Assets/Scripts/Factions/FactionActions.cs
@@ -5,7 +5,7 @@ using Game.Factions;
 
 public abstract class FactionActions
 {
-	public static void DoNothing(Faction faction)
+	public static void DoNothing(OldFaction faction)
 	{
 		OutputLogger.LogFormat("{0} Faction did nothing interesting during this cycle.", Game.Enums.LogSource.FACTIONACTION, faction.Name);
 	}
@@ -23,7 +23,7 @@ public abstract class MercantileActions
 
 public abstract class PoliticalActions
 {
-	public static void GenerateInfluence(Faction faction)
+	public static void GenerateInfluence(OldFaction faction)
 	{
 		//Replace this with an actual calculation based on faction factors
 		faction.influence++;
