+++ Assets/Scripts/Factions/OldFaction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7fdf14208ae954d468d2d7cd89fa52bb
+guid: 7b1ee55de28469f45b16620ffe57ab5c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
