+++ Assets/Scripts/SimulationManager.cs
@@ -152,7 +152,7 @@ public class SimulationManager : MonoBehaviour
 		}
         cityMarkers.Clear();
 
-        foreach(Faction faction in world.factions)
+        foreach(OldFaction faction in world.factions)
 		{
             foreach(City city in faction.cities)
 			{
